test: restore and split report tests
